+++ src/socket_test.ts
@@ -6,6 +6,22 @@ import { Ipv6Addr } from './ipv6.ts'
 import { SocketAddrV4 } from './socket.ts'
 import { SocketAddrV6 } from './socket.ts'
 
+Deno.test('socket address v4: constructor does not validate port number (too small)', () => {
+	const socket = new SocketAddrV4(
+		Ipv4Addr.newAddr(127, 0, 0, 1),
+		-1,
+	)
+	assertEquals(socket.port, -1)
+})
+
+Deno.test('socket address v4: constructor does not validate port number (too big)', () => {
+	const socket = new SocketAddrV4(
+		Ipv4Addr.newAddr(127, 0, 0, 1),
+		65536,
+	)
+	assertEquals(socket.port, 65536)
+})
+
 Deno.test('socket address v4: tryNew is ok', () => {
 	const socket = SocketAddrV4.tryNew(
 		Ipv4Addr.newAddr(127, 0, 0, 1),
@@ -38,7 +54,23 @@ Deno.test('socket address v4: tryNew errors if port is -Infinity', () => {
 	assertEquals(socket, null)
 })
 
-Deno.test('socket addresss v4: parse is ok', () => {
+Deno.test('socket address v4: tryNew errors if port is less than 0', () => {
+	const socket = SocketAddrV4.tryNew(
+		Ipv4Addr.newAddr(127, 0, 0, 1),
+		-1,
+	)
+	assertEquals(socket, null)
+})
+
+Deno.test('socket address v4: tryNew errors if port is greater than 65,535', () => {
+	const socket = SocketAddrV4.tryNew(
+		Ipv4Addr.newAddr(127, 0, 0, 1),
+		65536,
+	)
+	assertEquals(socket, null)
+})
+
+Deno.test('socket address v4: parse is ok', () => {
 	const socket = SocketAddrV4.parse('127.0.0.1:8080')
 	assert(socket instanceof SocketAddrV4)
 	assert(socket.addr.equals(Ipv4Addr.LOCALHOST))
@@ -75,6 +107,27 @@ Deno.test('socket address v4: parse is error, port is out of bounds', () => {
 	assertEquals(socket, null)
 })
 
+Deno.test('socket address v4: to string', () => {
+	const socket = SocketAddrV4.parse('0.0.0.0:8080') as SocketAddrV4
+	assertEquals(socket.toString(), '0.0.0.0:8080')
+})
+
+Deno.test('socket address v6: constructor does not validate port number (too small)', () => {
+	const socket = new SocketAddrV6(
+		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
+		-1,
+	)
+	assertEquals(socket.port, -1)
+})
+
+Deno.test('socket address v6: constructor does not validate port number (too big)', () => {
+	const socket = new SocketAddrV6(
+		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
+		65536,
+	)
+	assertEquals(socket.port, 65536)
+})
+
 Deno.test('socket address v6: tryNew is ok', () => {
 	const socket = SocketAddrV6.tryNew(
 		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
@@ -103,6 +156,22 @@ Deno.test('socket address v6: tryNew errors if port is -Infinity', () => {
 	const socket = SocketAddrV6.tryNew(
 		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
 		Number.NEGATIVE_INFINITY,
+	)
+	assertEquals(socket, null)
+})
+
+Deno.test('socket address v6: tryNew errors if port is less than 0', () => {
+	const socket = SocketAddrV6.tryNew(
+		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
+		-1,
+	)
+	assertEquals(socket, null)
+})
+
+Deno.test('socket address v6: tryNew errors if port is greater than 65,535', () => {
+	const socket = SocketAddrV6.tryNew(
+		Ipv6Addr.newAddr(1, 2, 3, 4, 5, 6, 7, 8),
+		65536,
 	)
 	assertEquals(socket, null)
 })